New time consistent clusters
In the previous version of the time consistent clusters, we used different cluster geometries for each year. These different cluster geometries were then used to compute population and worker counts for each cluster. This created complications with the data. Essentially, it amplified coding errors. For example, in some years, some parts of the cluster geometries would disappear (because they went below the density threshold). This in turn created big swings in population, largely dictated by the arbitrary threshold.

In the new version, we define the cluster geometry as the union of all cluster geometries throughout the years. We use this fixed geometry to compute population and other statistics for each year. This should ensure more consistency from one year to the next.